use Site scope rather than Data for getting speakers for schedule
The schedule partials do filtering of pages to get speaker info for the
schedule.  When the schedule is on the main page, using .Data.Pages
works fine, since all pages are children of the toplevel.  But on
/schedule that doesn't work, so use .Site.Pages instead.

Fixes the problem where /schedule only showed all the breaks, but none
of the actual speaker sessions.