Remove loader modal box from index.html
The loader logic was stalling until _all_ the content was loaded ... yup
every, single, image. Combined with enforced https which means nothing
is cached on the client it meant the loader would obscure the index page
for 30 seconds.

This must be killing our views and ticket sales.

You can see a sample of what this looks like at http://gophercon9000.cheney.net/