Add RSVP info for the Kickoff and After Parties
What
===
Add RSVP information for the Kickoff and After Parties.

Why
===
The Kickoff Party is hosted by the Denver Gophers Meetup and according to
https://twitter.com/GopherCon/status/746366250658439168 we need to RSVP
for the event, regardless of whether we have GopherCon tickets. This
isn't clear on the event page, and this change makes it explicit.

According to https://twitter.com/GopherCon/status/746373067400699904 the
After Party does not have an RSVP. Adding this section to this event
makes this clear.